leetCode 190

+++ src/main/scala/leetCode/Solution_190.scala
@@ -1,7 +1,7 @@
 package leetCode
 
 object Solution_190 {
-  def reverseBits(n: BigInt): Int = {
+  def reverseBits(n: Int): Int = {
     var res = 0
     (0 until 32).foreach(i => res |= ((n >> i) & 1) << (32 - i))
     res
